requirements.txt
numpy
tensorboardX
matplotlib
pyyaml